cardforge/requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0.6
django-admin==1.2.6
django-excel-response2==2.0.8
django-six==1.0.4
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
pkg-resources==0.0.0
psycopg2==2.7.5
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
screen==1.0.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
uritemplate==3.0.0
urllib3==1.23
xlwt==1.3.0
